@@ -12,5 +12,5 @@ use Illuminate\Foundation\Auth\Access\AuthorizesResources;
  */
 class Controller extends BaseController
 {
-    use AuthorizesRequests, AuthorizesResources, DispatchesJobs, ValidatesRequests;
+    use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
 }
