@@ -1,5 +1,16 @@
 <?php namespace App\Providers;
-
+/**
+ * Copyright 2017 OpenStack Foundation
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
 use App\Events\MyFavoritesAdd;
 use App\Events\SummitEventCreated;
 use App\Events\SummitEventDeleted;
@@ -7,13 +18,17 @@ use App\Events\SummitEventUpdated;
 use Illuminate\Contracts\Events\Dispatcher as DispatcherContract;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\App;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Event;
 use LaravelDoctrine\ORM\Facades\Registry;
+use models\main\AssetsSyncRequest;
 use models\summit\SummitEntityEvent;
 use App\Events\MyScheduleAdd;
 use App\Events\MyScheduleRemove;
 use Doctrine\ORM\Event\PreUpdateEventArgs;
 use models\utils\PreRemoveEventArgs;
+use IDCT\Networking\Ssh\SftpClient;
+use IDCT\Networking\Ssh\Credentials;
 
 /**
  * Class EventServiceProvider
@@ -35,12 +50,11 @@ class EventServiceProvider extends ServiceProvider
 
     /**
      * Register any other events for your application.
-     * @param  \Illuminate\Contracts\Events\Dispatcher $events
      * @return void
      */
-    public function boot(DispatcherContract $events)
+    public function boot()
     {
-        parent::boot($events);
+        parent::boot();
 
         Event::listen(\App\Events\MyScheduleAdd::class, function($event)
         {
@@ -272,6 +286,40 @@ class EventServiceProvider extends ServiceProvider
 
             $em = Registry::getManager('ss');
             $em->persist($entity_event);
+            $em->flush();
+
+        });
+
+        Event::listen(\App\Events\FileCreated::class, function($event)
+        {
+
+            $storage_path      = storage_path();
+            $local_path        = $event->getLocalPath();
+            $file_name         = $event->getFileName();
+            $folder_name       = $event->getFolderName();
+            $remote_base_path  = Config::get('scp.scp_remote_base_path', null);
+            $client            = new SftpClient();
+            $host              = Config::get('scp.scp_host', null);
+
+            $credentials       = Credentials::withPublicKey
+            (
+                Config::get('scp.ssh_user', null),
+                Config::get('scp.ssh_public_key', null),
+                Config::get('scp.ssh_private_key', null)
+            );
+
+            $client->setCredentials($credentials);
+            $client->connect($host);
+            $remote_destination = sprintf("%s/%s",$remote_base_path, $file_name);
+            $client->scpUpload(sprintf("%s/app/%s", $storage_path, $local_path), $remote_destination);
+            $client->close();
+
+            $asset_sync_request = new AssetsSyncRequest();
+            $asset_sync_request->setFrom($remote_destination);
+            $asset_sync_request->setTo(sprintf("%s/%s", $folder_name, $file_name));
+            $asset_sync_request->setProcessed(false);
+            $em = Registry::getManager('ss');
+            $em->persist($asset_sync_request);
             $em->flush();
 
         });
